+++ apps/io-sign-backoffice-func/src/infra/handlers/__test__/on-selfcare-contracts-message.spec.ts
@@ -1,13 +1,24 @@
-import { it, describe, expect, vi, beforeEach } from "vitest";
+import { it, describe, expect, vi } from "vitest";
 import * as TE from "fp-ts/lib/TaskEither";
 import { onSelfcareContractsMessageHandler } from "../on-selfcare-contracts-message";
 import { ioSignContracts } from "@/infra/selfcare/contract";
-import * as issuerMessage from "@/infra/slack/issuer-message";
 import { IoTsType } from "../validation";
 import { isLeft } from "fp-ts/lib/Either";
-import * as O from "fp-ts/lib/Option";
-import { GetById, Issuer } from "@/infra/back-office/issuer";
-import { SendMessage } from "@/infra/slack/message";
+
+import { User } from "@io-sign/io-sign/institution";
+import { Issuer, IssuerKey, IssuerRepository } from "@/issuer";
+
+import { google } from "googleapis";
+import { UserRepository } from "@/user";
+
+const auth = new google.auth.GoogleAuth({
+  credentials: {
+    type: "service_account",
+    private_key: "private_key",
+    client_email: "client_email",
+  },
+  scopes: ["https://www.googleapis.com/auth/spreadsheets"],
+});
 
 const issuer: Issuer = {
   id: "id",
@@ -18,32 +29,62 @@ const issuer: Issuer = {
   status: "active",
 };
 
-const mocks = { issuer };
+const users: User[] = [
+  {
+    name: "Nome",
+    surname: "Cognome",
+    email: "nome.cognome@contatto.pagopa.it",
+  },
+];
 
-type FunctionsTestContext = {
-  getById: GetById["getById"];
-  sendMessage: SendMessage["sendMessage"];
+const mocks = { issuer, users };
+
+const logger = {
+  log: () => () => {},
 };
 
-beforeEach<FunctionsTestContext>((ctx) => {
-  ctx.getById = (id, institutionId) =>
-    mocks.issuer.id === id && mocks.issuer.institutionId === institutionId
-      ? TE.right(O.some(issuer))
-      : TE.right(O.none);
-  ctx.sendMessage = () => TE.right(undefined);
-});
+const testRepository: IssuerRepository & UserRepository = {
+  getIssuerByKey: async (k: IssuerKey) => {
+    return k.id === mocks.issuer.id &&
+      k.institutionId === mocks.issuer.institutionId
+      ? mocks.issuer
+      : undefined;
+  },
+  getUsersByInstitutionId: async () => {
+    return mocks.users;
+  },
+};
+
+const mocksaveUsersToSpreadsheet = vi.hoisted(() =>
+  vi.fn(() => vi.fn(() => TE.right(undefined)))
+);
+
+vi.mock("@/infra/google/sheets", () => ({
+  saveUsersToSpreadsheet: mocksaveUsersToSpreadsheet,
+}));
+
+const sendMessageMock = vi.hoisted(() =>
+  vi.fn(() => vi.fn(() => TE.right(undefined)))
+);
+
+vi.mock("@/infra/slack/message", () => ({
+  sendMessage: sendMessageMock,
+}));
 
 describe("onSelfcareContractsMessage handler", () => {
-  it<FunctionsTestContext>("should return a left either when the input validation fails", ({
-    getById,
-    sendMessage,
-  }) => {
+  it("should return a left either when the input validation fails", () => {
     const run = onSelfcareContractsMessageHandler({
       inputDecoder: IoTsType(ioSignContracts),
       input: { foo: "foo" },
-      logger: { log: (s, _l) => () => console.log(s) },
-      getById,
-      sendMessage,
+      logger,
+      google: {
+        auth,
+        spreadsheetId: "",
+      },
+      userRepository: testRepository,
+      issuerRepository: testRepository,
+
+      slackWebhook: "https://my-web-hook",
     });
     expect(run()).resolves.toEqual(
       expect.objectContaining({
@@ -52,10 +93,7 @@ describe("onSelfcareContractsMessage handler", () => {
     );
   });
 
-  it<FunctionsTestContext>("should return a left either when the issuer already exists", async ({
-    getById,
-    sendMessage,
-  }) => {
+  it("should return a left either when the issuer already exists", async () => {
     const input = [
       {
         id: "id",
@@ -76,9 +114,14 @@ describe("onSelfcareContractsMessage handler", () => {
     const run = onSelfcareContractsMessageHandler({
       inputDecoder: IoTsType(ioSignContracts),
       input,
-      logger: { log: (s, _l) => () => console.log(s) },
-      getById,
-      sendMessage,
+      logger,
+      google: {
+        auth,
+        spreadsheetId: "",
+      },
+      userRepository: testRepository,
+      issuerRepository: testRepository,
+      slackWebhook: "https://my-web-hook",
     });
     const result = await run();
     expect(isLeft(result)).toBe(true);
@@ -88,11 +131,7 @@ describe("onSelfcareContractsMessage handler", () => {
     }
   });
 
-  it<FunctionsTestContext>("should call IssuerMessage function", async ({
-    getById,
-    sendMessage,
-  }) => {
-    const IssuerMessageSpy = vi.spyOn(issuerMessage, "IssuerMessage");
+  it("sends two messages (onboarding, contact) and stores contacts to spreadsheet", async () => {
     const input = [
       {
         id: "id",
@@ -100,7 +139,7 @@ describe("onSelfcareContractsMessage handler", () => {
         state: "ACTIVE",
         institution: {
           address: "address",
-          description: "description",
+          description: "Comune di Test",
           digitalAddress: "digitalAddress",
           taxCode: "taxCode",
         },
@@ -113,11 +152,20 @@ describe("onSelfcareContractsMessage handler", () => {
     const run = onSelfcareContractsMessageHandler({
       inputDecoder: IoTsType(ioSignContracts),
       input,
-      logger: { log: (s, _l) => () => console.log(s) },
-      getById,
-      sendMessage,
+      logger,
+      google: {
+        auth,
+        spreadsheetId: "",
+      },
+      userRepository: testRepository,
+      issuerRepository: testRepository,
+      slackWebhook: "https://my-web-hook",
     });
     await run();
-    expect(IssuerMessageSpy).toHaveBeenCalled();
+    expect(sendMessageMock).toHaveBeenCalledTimes(2);
+    expect(mocksaveUsersToSpreadsheet).toHaveBeenCalledWith(
+      mocks.users,
+      input[0].institution.description
+    );
   });
 });