+++ apps/io-sign-backoffice-func/src/infra/slack/message.ts
@@ -1,18 +1,14 @@
 import * as E from "fp-ts/lib/Either";
 import * as TE from "fp-ts/lib/TaskEither";
 import { pipe } from "fp-ts/lib/function";
-import { isSuccessful } from "@io-sign/io-sign/infra/client-utils";
+import { fetch } from "undici";
 import { dispatcher } from "../http/fetch";
 
-export type SendMessage = {
-  sendMessage: (message: string) => TE.TaskEither<Error, void>;
-};
-
-export const sendMessage = (webhookUrl: string) => (message: string) =>
+export const sendMessage = (message: string) => (r: { slackWebhook: string }) =>
   pipe(
     TE.tryCatch(
       () =>
-        fetch(webhookUrl, {
+        fetch(r.slackWebhook, {
           method: "POST",
           body: JSON.stringify({
             text: message,
@@ -22,7 +18,7 @@ export const sendMessage = (webhookUrl: string) => (message: string) =>
       E.toError
     ),
     TE.filterOrElse(
-      isSuccessful,
+      (response) => response.status === 200,
       () => new Error("The attempt to post message on slack failed.")
     ),
     TE.map(() => undefined)