+++ apps/io-sign-backoffice-func/CHANGELOG.md
@@ -1,5 +1,22 @@
 # io-sign-backoffice-func
 
+## 1.2.0
+
+### Minor Changes
+
+- bdc1236: Removed public API endpoints from backoffice-app
+
+### Patch Changes
+
+- Updated dependencies [bdc1236]
+  - @io-sign/io-sign@1.2.0
+
+## 1.1.0
+
+### Minor Changes
+
+- 5302648: Add integration with Google Spreadsheets
+
 ## 1.0.0
 
 ### Major Changes